fix: handle duplicate payment status updates and convert avgTipPercentage to double

+++ app/src/main/java/com/avoqado/pos/core/data/network/models/transactions/NetworkShiftRecord.kt
@@ -40,7 +40,7 @@ data class NetworkShiftRecord(
     @SerializedName("venueId")
     val venueId: String?,
     @SerializedName("avgTipPercentage")
-    val avgTipPercentage: Int?,
+    val avgTipPercentage: Double?,
     @SerializedName("tipsSum")
     val tipsSum: Int?,
     @SerializedName("tipsCount")
